# docs/swagger.yaml
basePath: /
definitions:
...
  dto.LoginRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      password:
        example: secret12345
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: a1b2c3d4e5f6...
        type: string
    type: object
  dto.LogoutRequest:
    properties:
      refresh_token:
        example: a1b2c3d4e5f6...
        type: string
    type: object
...
  dto.PaginatedRefreshTokenResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.RefreshTokenResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.RefreshRequest:
    properties:
      refresh_token:
        example: a1b2c3d4e5f6...
        type: string
    required:
    - refresh_token
    type: object
  dto.RefreshTokenResponse:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      expires_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      id:
        example: a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d
        type: string
      ip_address:
        example: 192.168.1.1
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      user_agent:
        example: Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML,
          like Gecko) Chrome/58.0.3029.110 Safari/537.3
        type: string
      user_id:
        example: a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      name:
        example: John Doe
        maxLength: 100
        minLength: 2
        type: string
      password:
        example: secret12345
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  dto.UserResponse:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d
        type: string
      name:
        example: John Doe
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
    type: object
...
  response.ApiResponse:
    properties:
      data: {}
      error:
        example: Error description
        type: string
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: Production-ready API boilerplate built with Gin, GORM, and Redis.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Gin Boilerplate API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, returning JWT access
        token and refresh token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with token pair
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "401":
          description: Unauthorized - Invalid email or password
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid payload validation
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: User authentication
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke access token by adding it to the blacklist and optionally
        revoke the refresh token
      parameters:
      - description: Optional refresh token to revoke alongside access token
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a newly issued access token
        and rotated refresh token
      parameters:
      - description: Refresh token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens refreshed successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "401":
          description: Unauthorized - Invalid, expired or revoked refresh token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid payload validation
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Refresh access token
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with name, email and password. Assigns
        default User role.
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "409":
          description: Conflict - Email already in use
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid payload validation
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Register a new user
      tags:
      - Auth
  /api/v1/auth/sessions:
    get:
      description: Retrieve paginated active refresh token sessions for the authenticated
        user, with optional filters and sorting
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Sorting criteria (e.g. created_at:desc, -expires_at)
        in: query
        name: sort
        type: string
      - description: Filter by User Agent (partial match)
        in: query
        name: user_agent
        type: string
      - description: Filter by IP Address (partial match)
        in: query
        name: ip_address
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sessions retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedRefreshTokenResponse'
              type: object
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid filter or sorting parameter
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List active user sessions
      tags:
      - Sessions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token (e.g. "Bearer eyJhbGci...").
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"